Selling functions created

+++ src/main/java/com/vitorgsevero/io/sistemavendas/repository/ClienteRepository.java
@@ -3,5 +3,5 @@ package com.vitorgsevero.io.sistemavendas.repository;
 import com.vitorgsevero.io.sistemavendas.model.Cliente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
+public interface ClienteRepository extends CrudRepository<Cliente, Long> {
 }
